Force Directed Schedule to cpp
From the original FORCE_DIRECTED.m I made the FORCE_DIRECTED_dot_cpp.m matlab code.  THen I ran the Matlab Coder function on that to generate a bunch of C++ files.

Instead of a cell array, the new output of the function is an array that has the cycle that each element should be scheduled on.